add dynamic border

+++ Notebooks/Transformer/data_flow.py
@@ -32,8 +32,8 @@ class Dataset_btc_hour(Dataset):
         self.scaler = StandardScaler()
         df_raw = pd.read_csv(self.data_path)
         len_ = len(df_raw)
-        border1s = [0, len_*0.8, len_*0.9]
-        border2s = [len_*0.8, len_*0.9, len_*1]
+        border1s = [0, int(len_*0.8), int(len_*0.9)]
+        border2s = [int(len_*0.8), int(len_*0.9), len_*1]
         border1 = border1s[self.set_type]
         border2 = border2s[self.set_type]
